chore: Add some tweakable variables for binary names
This is slightly nicer to folks building on other platforms, without having
to sacrifice anything on Linux.  With FreeBSD in mind, in particular, the
default `make` is bmake and luarocks binaries are typically named according
to the lua version they were built against -- having an easy way to avoid
patching in `gmake` and `luarocks54` would be quite welcome.